fix: update imports for moved targets
