+++ src/student_success_tool/preprocessing/targets/pdp/credits_earned.py
@@ -3,7 +3,7 @@ import typing as t
 import numpy as np
 import pandas as pd
 
-from ... import utils
+from .... import utils
 from . import shared
