+++ src/student_success_tool/preprocessing/targets/pdp/graduation.py
@@ -1,7 +1,7 @@
 import numpy as np
 import pandas as pd
 
-from ... import utils
+from .... import utils
 
 
 def compute_target(
