+++ src/student_success_tool/preprocessing/targets/pdp/retention.py
@@ -1,6 +1,6 @@
 import pandas as pd
 
-from ... import utils
+from .... import utils
 
 
 def compute_target(
