+++ tests/preprocessing/selection/test_pdp.py
@@ -4,7 +4,7 @@ import pandas as pd
 import pytest
 
 from student_success_tool.preprocessing.selection import pdp
-from student_success_tool.targets.pdp import shared
+from student_success_tool.preprocessing.targets.pdp import shared
 
 
 @pytest.fixture(scope="module")
