+++ tests/preprocessing/targets/pdp/test_graduation.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import pytest
 
-from student_success_tool.targets.pdp import graduation
+from student_success_tool.preprocessing.targets.pdp import graduation
 
 
 @pytest.mark.parametrize(
