+++ tests/preprocessing/targets/pdp/test_shared.py
@@ -1,7 +1,7 @@
 import pandas as pd
 import pytest
 
-from student_success_tool.targets.pdp import shared
+from student_success_tool.preprocessing.targets.pdp import shared
 
 
 @pytest.fixture(scope="module")
